feat(formsmanager): moved options into separate file
