+++ angular/playground/projects/ng-forms-manager/src/lib/forms-manager-options.ts
@@ -1,0 +1,13 @@
+import { InjectionToken } from '@angular/core';
+
+export type FormsManagerOptions = {
+  debounceTime: number;
+  updateStoreOnUnsubscribe: boolean;
+};
+
+export const FORMS_MANAGER_OPTIONS = new InjectionToken<FormsManagerOptions>('FormsManagerOptions');
+
+export const defaultOptions: FormsManagerOptions = {
+  debounceTime: 300,
+  updateStoreOnUnsubscribe: false
+};
