+++ angular/playground/projects/ng-forms-manager/src/public-api.ts
@@ -1,3 +1,4 @@
 export * from './lib/forms-manager';
+export { FORMS_MANAGER_OPTIONS, FormsManagerOptions } from './lib/forms-manager-options';
 export * from './lib/forms-manager.store';
 export * from './lib/forms-manager.query';
